Added dynamic chunk generation
Added new chunks generating around the player. Can still be improved by using partitioning to move the active chunks to the front of the vector of chunks.

+++ minecraft_clone/chunk.h
@@ -14,7 +14,7 @@
 struct Chunk
 {
 	glm::ivec2 coord{};
-	glm::vec2 world_position{};
+	glm::ivec2 world_position{};
 
 	bool is_active{ false };
